+++ CHANGELOG.md
@@ -2,8 +2,11 @@ This document record significant changes to the project
 
 ### 0.6.0 2024-01-12
 
-- Compile with java 17
-- Switch to using parent pom from rspace-os-parent project
+- compile with java 17
+- switch to using parent pom from rspace-os-parent project
+- remove `FigshareCategory.UNCATEGORIZED` static example category, as there is no such category in latest Figshare API. 
+For testing, use newly added `FigshareCategory.SOFTWARE_TESTING`. 
+ 
 
 ### 0.5.0 2022-06-08